chore: Copy correct API spec into repo
Currently, this repository contained the reworked API. That means that
messages like registerEGK were adjusted too. Now, it was updated to
reflect the currently implemented API. The API provided by the Gematik,
the API with additional messages and the reworked API can be found in
the repository: https://github.com/eHealthCardLink/api-ehcl-proposal
(see branches).